black reformat

+++ kblocks/extras/losses/lovasz/original.py
@@ -41,8 +41,8 @@ def lovasz_grad(gt_sorted):
     """
     gts = tf.reduce_sum(gt_sorted)
     intersection = gts - tf.cumsum(gt_sorted)
-    union = gts + tf.cumsum(1. - gt_sorted)
-    jaccard = 1. - intersection / union
+    union = gts + tf.cumsum(1.0 - gt_sorted)
+    jaccard = 1.0 - intersection / union
     jaccard = tf.concat((jaccard[0:1], jaccard[1:] - jaccard[:-1]), 0)
     return jaccard
 
@@ -83,25 +83,26 @@ def lovasz_hinge_flat(logits, labels):
 
     def compute_loss():
         labelsf = tf.cast(labels, logits.dtype)
-        signs = 2. * labelsf - 1.
-        errors = 1. - logits * tf.stop_gradient(signs)
-        errors_sorted, perm = tf.nn.top_k(errors,
-                                          k=tf.shape(errors)[0],
-                                          name="descending_sort")
+        signs = 2.0 * labelsf - 1.0
+        errors = 1.0 - logits * tf.stop_gradient(signs)
+        errors_sorted, perm = tf.nn.top_k(
+            errors, k=tf.shape(errors)[0], name="descending_sort"
+        )
         gt_sorted = tf.gather(labelsf, perm)
         grad = lovasz_grad(gt_sorted)
-        loss = tf.tensordot(tf.nn.relu(errors_sorted),
-                            tf.stop_gradient(grad),
-                            1,
-                            name="loss_non_void")
+        loss = tf.tensordot(
+            tf.nn.relu(errors_sorted), tf.stop_gradient(grad), 1, name="loss_non_void"
+        )
         return loss
 
     # deal with the void prediction case (only void pixels)
-    loss = tf.cond(tf.equal(tf.shape(logits)[0], 0),
-                   lambda: tf.reduce_sum(logits) * 0.,
-                   compute_loss,
-                   strict=True,
-                   name="loss")
+    loss = tf.cond(
+        tf.equal(tf.shape(logits)[0], 0),
+        lambda: tf.reduce_sum(logits) * 0.0,
+        compute_loss,
+        strict=True,
+        name="loss",
+    )
     return loss
 
 
@@ -115,20 +116,17 @@ def flatten_binary_scores(scores, labels, ignore=None):
     if ignore is None:
         return scores, labels
     valid = tf.not_equal(labels, ignore)
-    vscores = tf.boolean_mask(scores, valid, name='valid_scores')
-    vlabels = tf.boolean_mask(labels, valid, name='valid_labels')
+    vscores = tf.boolean_mask(scores, valid, name="valid_scores")
+    vlabels = tf.boolean_mask(labels, valid, name="valid_labels")
     return vscores, vlabels
 
 
 # --------------------------- MULTICLASS LOSSES ---------------------------
 
 
-def lovasz_softmax(probas,
-                   labels,
-                   classes='present',
-                   per_image=False,
-                   ignore=None,
-                   order='BHWC'):
+def lovasz_softmax(
+    probas, labels, classes="present", per_image=False, ignore=None, order="BHWC"
+):
     """
     Multi-class Lovasz-Softmax loss
       probas: [B, H, W, C] or [B, C, H, W] Variable, class probabilities at each prediction (between 0 and 1)
@@ -151,11 +149,12 @@ def lovasz_softmax(probas,
         loss = tf.reduce_mean(losses)
     else:
         loss = lovasz_softmax_flat(
-            *flatten_probas(probas, labels, ignore, order), classes=classes)
+            *flatten_probas(probas, labels, ignore, order), classes=classes
+        )
     return loss
 
 
-def lovasz_softmax_flat(probas, labels, classes='present'):
+def lovasz_softmax_flat(probas, labels, classes="present"):
     """
     Multi-class Lovasz-Softmax loss
       probas: [P, C] Variable, class probabilities at each prediction (between 0 and 1)
@@ -165,56 +164,55 @@ def lovasz_softmax_flat(probas, labels, classes='present'):
     C = probas.shape[1]
     losses = []
     present = []
-    class_to_sum = list(range(C)) if classes in ['all', 'present'] else classes
+    class_to_sum = list(range(C)) if classes in ["all", "present"] else classes
     for c in class_to_sum:
-        fg = tf.cast(tf.equal(labels, c),
-                     probas.dtype)  # foreground for class c
-        if classes == 'present':
+        fg = tf.cast(tf.equal(labels, c), probas.dtype)  # foreground for class c
+        if classes == "present":
             present.append(tf.reduce_sum(fg) > 0)
         if C == 1:
             if len(classes) > 1:
-                raise ValueError('Sigmoid output possible only with 1 class')
+                raise ValueError("Sigmoid output possible only with 1 class")
             class_pred = probas[:, 0]
         else:
             class_pred = probas[:, c]
         errors = tf.abs(fg - class_pred)
-        errors_sorted, perm = tf.nn.top_k(errors,
-                                          k=tf.shape(errors)[0],
-                                          name="descending_sort_{}".format(c))
+        errors_sorted, perm = tf.nn.top_k(
+            errors, k=tf.shape(errors)[0], name="descending_sort_{}".format(c)
+        )
         fg_sorted = tf.gather(fg, perm)
         grad = lovasz_grad(fg_sorted)
         losses.append(
-            tf.tensordot(errors_sorted,
-                         tf.stop_gradient(grad),
-                         1,
-                         name="loss_class_{}".format(c)))
+            tf.tensordot(
+                errors_sorted, tf.stop_gradient(grad), 1, name="loss_class_{}".format(c)
+            )
+        )
     if len(class_to_sum) == 1:  # short-circuit mean when only one class
         return losses[0]
     losses_tensor = tf.stack(losses)
-    if classes == 'present':
+    if classes == "present":
         present = tf.stack(present)
         losses_tensor = tf.boolean_mask(losses_tensor, present)
     loss = tf.reduce_mean(losses_tensor)
     return loss
 
 
-def flatten_probas(probas, labels, ignore=None, order='BHWC'):
+def flatten_probas(probas, labels, ignore=None, order="BHWC"):
     """
     Flattens predictions in the batch
     """
     if len(probas.shape) == 3:
-        probas, order = tf.expand_dims(probas, 3), 'BHWC'
-    if order == 'BCHW':
+        probas, order = tf.expand_dims(probas, 3), "BHWC"
+    if order == "BCHW":
         probas = tf.transpose(probas, (0, 2, 3, 1), name="BCHW_to_BHWC")
-        order = 'BHWC'
-    if order != 'BHWC':
-        raise NotImplementedError('Order {} unknown'.format(order))
+        order = "BHWC"
+    if order != "BHWC":
+        raise NotImplementedError("Order {} unknown".format(order))
     C = probas.shape[3]
     probas = tf.reshape(probas, (-1, C))
     labels = tf.reshape(labels, (-1,))
     if ignore is None:
         return probas, labels
     valid = tf.not_equal(labels, ignore)
-    vprobas = tf.boolean_mask(probas, valid, name='valid_probas')
-    vlabels = tf.boolean_mask(labels, valid, name='valid_labels')
+    vprobas = tf.boolean_mask(probas, valid, name="valid_probas")
+    vlabels = tf.boolean_mask(labels, valid, name="valid_labels")
     return vprobas, vlabels
